Use xenial in travis ci(as older versions seem to have problems with

+++ install-opengl.ps1
@@ -4,7 +4,7 @@
 # Original script: https://github.com/glue-viz/glue-vispy-viewers/blob/master/.install-opengl.ps1
 
 #$MESA_GL_URL = "https://github.com/vispy/demo-data/raw/master/mesa/"
-$MESA_GL_URL = "http://downloads.fdossena.com/Projects/Mesa3D/Builds/MesaForWindows-17.3.3.zip"
+$MESA_GL_URL = "https://ci.appveyor.com/api/projects/jvbsl/mesabinary/artifacts/opengl_win_dll.zip?branch=master"
 # Mesa DLLs found linked from:
 #     http://qt-project.org/wiki/Cross-compiling-Mesa-for-Windows
 # to:
@@ -19,25 +19,17 @@ function revokeRights($file) {
         icacls "$file" /inheritance:d
         icacls "$file" /setowner "NT SERVICE\TrustedInstaller"
 }
-
-function DownloadMesaOpenGL ($architecture) {
+function DownloadMesaOpenGL() {
     $webclient = New-Object System.Net.WebClient
-    $basedir = "$env:WINDIR\system32\"
     $filepath = $basedir + "opengl32.dll"
     # Download and retry up to 3 times in case of network transient errors.
     # $url = $MESA_GL_URL + "opengl32_mingw_" + $architecture + ".dll"
     $url = $MESA_GL_URL
-
-    if (Test-Path $filepath) {
-        Write-Host "change old opengl32.dll rights"
-        grantRights $filepath
-        Write-Host "Move old opengl32.dll(->opengl32.dll_old) in favour of mesa opengl"
-        Rename-Item -Path $filepath -NewName opengl32.dll_old
-    }
-    If(!(test-path "./temp"))
+    If (test-path "./temp")
     {
-          New-Item -ItemType Directory -Force -Path "./temp"
+        Remove-Item -Recurse -Force -Path "./temp"
     }
+    New-Item -ItemType Directory -Force -Path "./temp"
     $temppath = "./temp/mesa.zip"
     Write-Host "Downloading" $url
     $retry_attempts = 2
@@ -45,8 +37,6 @@ function DownloadMesaOpenGL ($architecture) {
         try {
             $webclient.DownloadFile($url, $temppath)
             Expand-Archive -Force $temppath ./temp/
-            Write-Host "Try saving at" $filepath
-            cp ./temp/opengl32.dll $filepath
             break
         }
         Catch [Exception]{
@@ -60,24 +50,70 @@ function DownloadMesaOpenGL ($architecture) {
         $webclient.DownloadFile($url, $temppath)
     }
 }
-
-function RemoveMesaOpenGL()
+function InstallMesaOpenGL ($arch,$basedir)
 {
-    $basedir = "$env:WINDIR/system32/"
+    $source_path = "./temp/build/$arch/opengl32.dll"
+    if (Test-Path $source_path)
+    {
+        $filepath = $basedir + "opengl32.dll"
+        if (Test-Path $filepath)
+        {
+            Write-Host "change old opengl32.dll rights"
+            grantRights $filepath
+            Write-Host "Move old opengl32.dll(->opengl32.dll_old) in favour of mesa opengl"
+            Rename-Item -Path $filepath -NewName opengl32.dll_old
+        }
+        Write-Host "Try saving at" $filepath
+        cp $source_path $filepath
+    }
+    else
+    {
+        Write-Host "OpenGL32 not found: $source_path"
+    }
+}
+function InstallAllMesaOpenGL()
+{
+    if ([IntPtr]::Size -eq 8)
+    {
+        InstallMesaOpenGL "x86" "$env:WINDIR/SysWOW64/"
+        InstallMesaOpenGL "x86_64" "$env:WINDIR/system32/"
+    }
+    elseif([IntPtr]::Size -eq 4)
+    {
+        InstallMesaOpenGL "x86" "$env:WINDIR/system32/"
+    }
+    else
+    {
+        Write-Host "Invalid host architecture"
+    }
+}
+
+function RemoveMesaOpenGL($basedir)
+{
     $filepath = $basedir + "opengl32.dll"
     $filepath_old = $basedir + "opengl32.dll_old"
-    Write-Host "Restore old opengl32.dll"
-    Remove-Item $filepath -Force
-    Rename-Item -Path $filepath_old -NewName opengl32.dll
-    revokeRights $filePath
+    if (Test-Path $filepath_old)
+    {
+        Write-Host "Restore old opengl32.dll"
+        Remove-Item $filepath -Force
+        Rename-Item -Path $filepath_old -NewName opengl32.dll
+        revokeRights $filePath
+    }
+}
+
+function RemoveAllMesaOpenGL()
+{
+    RemoveMesaOpenGL "$env:WINDIR/system32/"
+    RemoveMesaOpenGL "$env:WINDIR/SysWOW64/"
 }
 
 
 if (($Args.Count -gt 0) -and ($args[0] = "uninstall"))
 {
-    RemoveMesaOpenGL
+    RemoveAllMesaOpenGL
 }
 else
 {
-    DownloadMesaOpenGL $env:PYTHON_ARCH
+    DownloadMesaOpenGL
+    InstallAllMesaOpenGL 
 }
